Initial checkin for Serial Matlab scripts
Adding .gitignore for matlab scripts
Adding SerialTest.m, which is a template for user interface over a serial
port in a matlab script.  This works with USB CDC and TestAudioTransfer
in the ST32M repo.
Adding SimpleSignalFFT.m.  This script generates a sample wavform, runs
FFT magnitude and spectrogram plots.  The script also creates a text file
with the sample waveform in int16 format for use in C programs.
Adding sampleaudio.txt.  This is the output from SimpleSignalFFT as an
example for users.